# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Gateway",
    "version": "1.0.0"
  },
  "servers": [],
  "paths": {
    "/auth/signup": {
      "post": {
        "summary": "Sign Up",
        "operationId": "signUp",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "format": "password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Signup successful"
          }
        },
        "servers": [
          {
            "url": "http://localhost:3002"
          }
        ]
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login",
        "operationId": "login",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "format": "password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful"
          }
        },
        "servers": [
          {
            "url": "http://localhost:8080/api"
          }
        ]
      }
    },
    "/shortened-url": {
      "post": {
        "summary": "Create Shortened URL",
        "operationId": "createShortenedUrl",
        "tags": ["Shortened URL"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["originalUrl"],
                "properties": {
                  "originalUrl": {
                    "type": "string",
                    "format": "url"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "URL shortened successfully"
          }
        },
        "servers": [
          {
            "url": "http://localhost:3001"
          }
        ]
      },
      "get": {
        "summary": "Get All Shortened URLs",
        "operationId": "getAllShortenedUrls",
        "tags": ["Shortened URL"],
        "responses": {
          "200": {
            "description": "List of shortened URLs"
          }
        },
        "servers": [
          {
            "url": "http://localhost:3001"
          }
        ]
      }
    },
    "/shortened-url/{shortCode}": {
      "get": {
        "summary": "Redirect to Original URL",
        "operationId": "redirectToOriginal",
        "tags": ["Shortened URL"],
        "parameters": [
          {
            "name": "shortCode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The short code associated with the original URL"
          }
        ],
        "responses": {
          "200": {
            "description": "Redirection successful",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string",
                  "example": "<html>...</html>"
                }
              }
            }
          },
          "404": {
            "description": "URL not found"
          }
        },
        "servers": [
          {
            "url": "http://localhost:3001"
          }
        ]
      }
    },
    "/shortened-url/{id}": {
      "patch": {
        "summary": "Update Shortened URL",
        "operationId": "updateShortenedUrl",
        "tags": ["Shortened URL"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["originalUrl"],
                "properties": {
                  "originalUrl": {
                    "type": "string",
                    "format": "url"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "URL updated successfully"
          }
        },
        "servers": [
          {
            "url": "http://localhost:3001"
          }
        ]
      },
      "delete": {
        "summary": "Delete Shortened URL",
        "operationId": "deleteShortenedUrl",
        "tags": ["Shortened URL"],
        "responses": {
          "200": {
            "description": "URL deleted successfully"
          }
        },
        "servers": [
          {
            "url": "http://localhost:3001"
          }
        ]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "tags": [
    {
      "name": "Auth",
      "description": "Operations related to authentication"
    },
    {
      "name": "Shortened URL",
      "description": "Operations related to URL shortening"
    }
  ]
}